Make sure to give an integer to ipythoncluster
Mainly pipeline affected for now chipseq.
I need to reduce the number of jobs to half of the samples,
so will divide by a float and need to convert to integer after
that.

+++ bcbio/chipseq/peaks.py
@@ -75,4 +75,4 @@ def _get_multiplier(samples):
     for sample in samples:
         if dd.get_phenotype(sample[0]) == "chip":
             to_process += 1.0
-    return int(to_process / len(samples))
+    return to_process / len(samples)
